PDFBOX-2147: Removed unneccesary unboxing
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/trunk@1609404 13f79535-47bb-0310-9956-ffa450edef68

+++ pdfbox/src/main/java/org/apache/pdfbox/pdmodel/PDPage.java
@@ -572,7 +572,7 @@ public class PDPage implements COSObjectable
         Integer rotation = getRotation();
         if (rotation != null)
         {
-            retval = rotation.intValue();
+            retval = rotation;
         }
         else
         {
